[feat] replace sync threadpool for deferred lock
This is a pending change synce we transformed the soledad sync into twisted
async style. The threadpool previously used for initializing the database and
for running the sync is not needed, so this commit removes all that and adds a
deferred lock to make sure all syncs are mutually exclusive.